Add files via upload
I didn't know that this was a byproduct of Fermat's little theorem. I just discovered it while tinkering with the AKS algorithm. This is essentially Fermat's primality test but I didn't know that when I made it because I'd never heard of it and I didn't know it could fail and produce "pseudoprimes". It doesn't fail up to 63 which is as high as I could test on my laptop. It's not very efficient in that sense but lightning fast if it didn't require more bits than there are stars in the universe beyond a certain point.